requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.49.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tornado==6.4
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
zipp==3.17.0